web: renaming for clarity
