+++ web/src/routeTree.gen.ts
@@ -17,8 +17,8 @@ import { Route as LegalPrivacyRouteImport } from './routes/legal/privacy'
 import { Route as AppAppLayoutRouteImport } from './routes/app/_appLayout'
 import { Route as AppAppLayoutIndexRouteImport } from './routes/app/_appLayout/index'
 import { Route as AppAppLayoutToolsRouteImport } from './routes/app/_appLayout/tools'
+import { Route as AppAppLayoutProvidersRouteImport } from './routes/app/_appLayout/providers'
 import { Route as AppAppLayoutProfileRouteImport } from './routes/app/_appLayout/profile'
-import { Route as AppAppLayoutAppKeysRouteImport } from './routes/app/_appLayout/app-keys'
 import { Route as AppAppLayoutApiKeysRouteImport } from './routes/app/_appLayout/api-keys'
 import { Route as AppAppLayoutSessionSessionIdRouteImport } from './routes/app/_appLayout/session/$sessionId'
 
@@ -58,14 +58,14 @@ const AppAppLayoutToolsRoute = AppAppLayoutToolsRouteImport.update({
   path: '/tools',
   getParentRoute: () => AppAppLayoutRoute,
 } as any)
+const AppAppLayoutProvidersRoute = AppAppLayoutProvidersRouteImport.update({
+  id: '/providers',
+  path: '/providers',
+  getParentRoute: () => AppAppLayoutRoute,
+} as any)
 const AppAppLayoutProfileRoute = AppAppLayoutProfileRouteImport.update({
   id: '/profile',
   path: '/profile',
-  getParentRoute: () => AppAppLayoutRoute,
-} as any)
-const AppAppLayoutAppKeysRoute = AppAppLayoutAppKeysRouteImport.update({
-  id: '/app-keys',
-  path: '/app-keys',
   getParentRoute: () => AppAppLayoutRoute,
 } as any)
 const AppAppLayoutApiKeysRoute = AppAppLayoutApiKeysRouteImport.update({
@@ -86,8 +86,8 @@ export interface FileRoutesByFullPath {
   '/legal/privacy': typeof LegalPrivacyRoute
   '/legal/terms': typeof LegalTermsRoute
   '/app/api-keys': typeof AppAppLayoutApiKeysRoute
-  '/app/app-keys': typeof AppAppLayoutAppKeysRoute
   '/app/profile': typeof AppAppLayoutProfileRoute
+  '/app/providers': typeof AppAppLayoutProvidersRoute
   '/app/tools': typeof AppAppLayoutToolsRoute
   '/app/': typeof AppAppLayoutIndexRoute
   '/app/session/$sessionId': typeof AppAppLayoutSessionSessionIdRoute
@@ -98,8 +98,8 @@ export interface FileRoutesByTo {
   '/legal/privacy': typeof LegalPrivacyRoute
   '/legal/terms': typeof LegalTermsRoute
   '/app/api-keys': typeof AppAppLayoutApiKeysRoute
-  '/app/app-keys': typeof AppAppLayoutAppKeysRoute
   '/app/profile': typeof AppAppLayoutProfileRoute
+  '/app/providers': typeof AppAppLayoutProvidersRoute
   '/app/tools': typeof AppAppLayoutToolsRoute
   '/app/session/$sessionId': typeof AppAppLayoutSessionSessionIdRoute
 }
@@ -111,8 +111,8 @@ export interface FileRoutesById {
   '/legal/privacy': typeof LegalPrivacyRoute
   '/legal/terms': typeof LegalTermsRoute
   '/app/_appLayout/api-keys': typeof AppAppLayoutApiKeysRoute
-  '/app/_appLayout/app-keys': typeof AppAppLayoutAppKeysRoute
   '/app/_appLayout/profile': typeof AppAppLayoutProfileRoute
+  '/app/_appLayout/providers': typeof AppAppLayoutProvidersRoute
   '/app/_appLayout/tools': typeof AppAppLayoutToolsRoute
   '/app/_appLayout/': typeof AppAppLayoutIndexRoute
   '/app/_appLayout/session/$sessionId': typeof AppAppLayoutSessionSessionIdRoute
@@ -125,8 +125,8 @@ export interface FileRouteTypes {
     | '/legal/privacy'
     | '/legal/terms'
     | '/app/api-keys'
-    | '/app/app-keys'
     | '/app/profile'
+    | '/app/providers'
     | '/app/tools'
     | '/app/'
     | '/app/session/$sessionId'
@@ -137,8 +137,8 @@ export interface FileRouteTypes {
     | '/legal/privacy'
     | '/legal/terms'
     | '/app/api-keys'
-    | '/app/app-keys'
     | '/app/profile'
+    | '/app/providers'
     | '/app/tools'
     | '/app/session/$sessionId'
   id:
@@ -149,8 +149,8 @@ export interface FileRouteTypes {
     | '/legal/privacy'
     | '/legal/terms'
     | '/app/_appLayout/api-keys'
-    | '/app/_appLayout/app-keys'
     | '/app/_appLayout/profile'
+    | '/app/_appLayout/providers'
     | '/app/_appLayout/tools'
     | '/app/_appLayout/'
     | '/app/_appLayout/session/$sessionId'
@@ -214,18 +214,18 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AppAppLayoutToolsRouteImport
       parentRoute: typeof AppAppLayoutRoute
     }
+    '/app/_appLayout/providers': {
+      id: '/app/_appLayout/providers'
+      path: '/providers'
+      fullPath: '/app/providers'
+      preLoaderRoute: typeof AppAppLayoutProvidersRouteImport
+      parentRoute: typeof AppAppLayoutRoute
+    }
     '/app/_appLayout/profile': {
       id: '/app/_appLayout/profile'
       path: '/profile'
       fullPath: '/app/profile'
       preLoaderRoute: typeof AppAppLayoutProfileRouteImport
-      parentRoute: typeof AppAppLayoutRoute
-    }
-    '/app/_appLayout/app-keys': {
-      id: '/app/_appLayout/app-keys'
-      path: '/app-keys'
-      fullPath: '/app/app-keys'
-      preLoaderRoute: typeof AppAppLayoutAppKeysRouteImport
       parentRoute: typeof AppAppLayoutRoute
     }
     '/app/_appLayout/api-keys': {
@@ -247,8 +247,8 @@ declare module '@tanstack/react-router' {
 
 interface AppAppLayoutRouteChildren {
   AppAppLayoutApiKeysRoute: typeof AppAppLayoutApiKeysRoute
-  AppAppLayoutAppKeysRoute: typeof AppAppLayoutAppKeysRoute
   AppAppLayoutProfileRoute: typeof AppAppLayoutProfileRoute
+  AppAppLayoutProvidersRoute: typeof AppAppLayoutProvidersRoute
   AppAppLayoutToolsRoute: typeof AppAppLayoutToolsRoute
   AppAppLayoutIndexRoute: typeof AppAppLayoutIndexRoute
   AppAppLayoutSessionSessionIdRoute: typeof AppAppLayoutSessionSessionIdRoute
@@ -256,8 +256,8 @@ interface AppAppLayoutRouteChildren {
 
 const AppAppLayoutRouteChildren: AppAppLayoutRouteChildren = {
   AppAppLayoutApiKeysRoute: AppAppLayoutApiKeysRoute,
-  AppAppLayoutAppKeysRoute: AppAppLayoutAppKeysRoute,
   AppAppLayoutProfileRoute: AppAppLayoutProfileRoute,
+  AppAppLayoutProvidersRoute: AppAppLayoutProvidersRoute,
   AppAppLayoutToolsRoute: AppAppLayoutToolsRoute,
   AppAppLayoutIndexRoute: AppAppLayoutIndexRoute,
   AppAppLayoutSessionSessionIdRoute: AppAppLayoutSessionSessionIdRoute,
