+++ web/src/routes/app/_appLayout/api-keys.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 
-import { ProviderKeysManager } from "@/components/ApiKeysManager";
+import { ApiKeysManager } from "@/components/ApiKeysManager";
 
 export const Route = createFileRoute("/app/_appLayout/api-keys")({
   component: RouteComponent,
@@ -10,7 +10,7 @@ function RouteComponent() {
   return (
     <div className="overflow-auto bg-background">
       <div className="container mx-auto px-4 py-8 max-w-2xl">
-        <ProviderKeysManager />
+        <ApiKeysManager />
       </div>
     </div>
   );
