+++ web/src/routes/app/_appLayout/providers.tsx
@@ -1,8 +1,8 @@
 import { createFileRoute } from "@tanstack/react-router";
 
-import { ApiKeysManager } from "@/components/AppApiKeysManager";
+import { ProviderManager } from "@/components/ProviderManager";
 
-export const Route = createFileRoute("/app/_appLayout/app-keys")({
+export const Route = createFileRoute("/app/_appLayout/providers")({
   component: RouteComponent,
 });
 
@@ -10,7 +10,7 @@ function RouteComponent() {
   return (
     <div className="overflow-auto bg-background">
       <div className="container mx-auto px-4 py-8 max-w-2xl">
-        <ApiKeysManager />
+        <ProviderManager />
       </div>
     </div>
   );
